Fixed home page
Made sure that in mobile page, when the dropdown happens everything
flows under the dropdown. Had to adjust for regular screens as a result.
Did this by position relative, and top/bottom adjustments. Also had to
add in JS code to control top of left/right carousel controls on click.